agrego testeos unitarios, login y manejo de excepciones

+++ src/main/java/com/dh/IntegradorBackend/service/DomicilioService.java
@@ -3,6 +3,8 @@ package com.dh.IntegradorBackend.service;
 
 
 import com.dh.IntegradorBackend.entities.Domicilio;
+import com.dh.IntegradorBackend.exceptions.BadRequestException;
+import com.dh.IntegradorBackend.exceptions.ResourceNotFoundException;
 import com.dh.IntegradorBackend.repository.DomicilioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,16 +29,31 @@ public class DomicilioService {
         return (Domicilio) repository.save(domicilio);
     }
 
-    public void eliminarDomicilio(Long id){
-        repository.deleteById(id);
+    public void eliminarDomicilio(Long id)throws ResourceNotFoundException{
+        Optional<Domicilio> domicilio = buscarDomicilio(id);
+        if (domicilio.isPresent())
+            repository.deleteById(id);
+        else
+            throw new ResourceNotFoundException("No existe el domicilio con el id" + id + ", no se pudo eliminar.");
     }
 
-    public Domicilio actualizarDomicilio(Domicilio domicilio){
+    public Domicilio actualizarDomicilio(Domicilio domicilio) throws BadRequestException {
         if (buscarDomicilio(domicilio.getId()).isPresent()){
             return (Domicilio) repository.save(domicilio);
         }
-        return null;
+        else {
+            throw new BadRequestException("No se encontro el domicilio, no fue actualizado");
+        }
     }
+
+    /*public Domicilio actualizarDomicilio(Domicilio domicilio) {
+        if (buscarDomicilio(domicilio.getId()).isPresent()){
+            return (Domicilio) repository.save(domicilio);
+        }
+        else {
+            return null;
+        }
+    }*/
 
 
 }
