+++ src/main/java/com/dh/IntegradorBackend/service/PacienteService.java
@@ -1,6 +1,8 @@
 package com.dh.IntegradorBackend.service;
 
 import com.dh.IntegradorBackend.entities.Paciente;
+import com.dh.IntegradorBackend.exceptions.BadRequestException;
+import com.dh.IntegradorBackend.exceptions.ResourceNotFoundException;
 import com.dh.IntegradorBackend.repository.PacienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,6 +21,7 @@ public class PacienteService {
     }
 
     public Optional<Paciente> buscarPaciente(Long id){
+        Optional<Paciente> paciente = repository.findById(id);
         return repository.findById(id);
     }
 
@@ -26,18 +29,44 @@ public class PacienteService {
         return repository.save(paciente);
     }
 
-    public Paciente actualizarPaciente(Paciente paciente){
+    public Paciente actualizarPaciente (Paciente paciente)throws BadRequestException{
+
         if (buscarPaciente(paciente.getId()).isPresent()){
             return (Paciente) repository.save(paciente);
+        }else{
+            throw new BadRequestException("No se encontro el paciente, no fue actualizado ");
         }
-        return null;
+
     }
 
-    public void eliminarPaciente(Long id){
-        repository.deleteById(id);
+    /*public Paciente actualizarPaciente (Paciente paciente){
+
+        if (buscarPaciente(paciente.getId()).isPresent()){
+            return (Paciente) repository.save(paciente);
+        }else{
+            return null;
+        }
+
+    }*/
+
+    public void eliminarPaciente(Long id) throws ResourceNotFoundException {
+        Optional<Paciente> paciente = buscarPaciente(id);
+        if(paciente.isPresent()){
+            repository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("No existe el paciente con el id" + id + ", no se pudo eliminar.");
+        }
+
     }
 
-    public Optional<Paciente> buscarPorEmail(String email){
-        return repository.buscarPacientePorEmail(email);
+    public Optional<Paciente> buscarPorEmail(String email) throws ResourceNotFoundException {
+        Optional<Paciente> paciente = repository.buscarPacientePorEmail(email);
+        if (paciente.isPresent()){
+            return paciente;
+        }else {
+            throw new ResourceNotFoundException("No se encontro el paciente con email " + email );
+        }
     }
+
+
 }
