+++ src/main/java/com/dh/IntegradorBackend/service/TurnoService.java
@@ -1,6 +1,8 @@
 package com.dh.IntegradorBackend.service;
 
 import com.dh.IntegradorBackend.entities.Turno;
+import com.dh.IntegradorBackend.exceptions.BadRequestException;
+import com.dh.IntegradorBackend.exceptions.ResourceNotFoundException;
 import com.dh.IntegradorBackend.repository.TurnoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,15 +28,30 @@ public class TurnoService {
         return (Turno) repository.save(turno);
     }
 
-    public Turno actualizarTurno(Turno turno){
+    public Turno actualizarTurno(Turno turno) throws BadRequestException {
         if (buscarTurno(turno.getId()).isPresent()){
             return (Turno) repository.save(turno);
         }
-        return null;
+        else {
+            throw new BadRequestException("No se encontro el turno, no fue actualizado");
+        }
     }
 
-    public void eliminarTurno(Long id){
-        repository.deleteById(id);
+    /*public Turno actualizarTurno(Turno turno) {
+        if (buscarTurno(turno.getId()).isPresent()){
+            return (Turno) repository.save(turno);
+        }
+        else {
+            return null;
+        }
+    }*/
+
+    public void eliminarTurno(Long id) throws ResourceNotFoundException {
+        Optional<Turno> turno = buscarTurno(id);
+        if (turno.isPresent())
+            repository.deleteById(id);
+        else
+            throw new ResourceNotFoundException("No existe el turno con el id" + id + ", no se pudo eliminar.");
     }
 
 }
